feat(rqf-21): Sequencia do Codigo da UP com 4 digitos
Esse commit faz as proximas Unidades de Producao cadastradas terem um codigo de sequencia de 4

digitos.

BREAKING CHANGE: A partir deste commit, o campo CodigoPropriedade da Classe UnidadeProducao precisa

de ser um Int64 e não mais um Int32, pois ele tem um tamanho maior que este. O campo atual no banco

é um Number(38, 0), incompatível com Int64. É necessário executar um alter table no banco caso a

tabela CRT_UNIDADE_PRODUCAO na coluna PROPRIEDADE_CODIGO não se comporte como deve.

+++ Código Fonte/VERSAO-ATUAL/lib/src/Blocos.Entities/Interno.Extensoes/Caracterizacoes/ModuloUnidadeProducao/UnidadeProducao.cs
@@ -11,7 +11,7 @@ namespace Tecnomapas.Blocos.Entities.Interno.Extensoes.Caracterizacoes.ModuloUni
 		public String Tid { get; set; }
 		public String LocalLivroDisponivel { get; set; }
 		public Boolean PossuiCodigoPropriedade { get; set; }
-		public Int32 CodigoPropriedade { get; set; }
+		public Int64 CodigoPropriedade { get; set; }
 		public int InternoID { get; set; }
 		public string InternoTID { get; set; }
 		public int CredenciadoID { get; set; }
